+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 ### Fixed
 - Subsonic API: In methods `getAlbumInfo` and `getAlbumInfo2` with no info found, an empty array was returned instead of an empty object
   [#1233](https://github.com/owncloud/music/issues/1233)
+- File hooks potentially breaking the cloud update process when Music app is updated at the same time and the cloud runs partially updated app instance
+  [#1231](https://github.com/owncloud/music/issues/1231)
 
 ## 2.2.0 - 2025-06-15